Added timeouts to wait for transmission loops
Hi, I added timeouts to wait for transmission loops to avoid the sketch getting stuck if the memory chip is not found of the bus fails for some reason, i.e. no pullups, broken wire, etc.

I used the existing MAX_ITER constant but hits might be changed depending on how long we need to wait.

As the timeout action I copied the action on the default: case of the corresponding switch statement. Please review if this is correct since I'm not a great specialist on how the I2C bus works.

I hope this works!