Add separate timeout + fixed while exit condition
Hi, thanks for having a look.

Following on your last comments, added a separate timeout so it is possible to have different wait times. Set to 5000 just by intuition, feel free to modify according to your experience.

Also fixed the while exit condition.

Regarding the counter for begin:... honestly no idea if that would be necessary, i'll let it to your own judgement if including it or not :)

I should also have an available board in some days so I should be able to run some tests myself.