Tests done
I run some tests with a Mini Pro 8Mhz and I2C memory breadboard.
I updated the timeout detection and added error codes so that which timeout was triggered can be determined.
If everything works ok, the serial device will output SSF2XA. If the first timeout is triggered, then the output will be SSF21XA; if it was the second timeout, it will be SSF22XA, and so on.

I run the bootloader after disconnecting the wires from the memory and it seems to work correctly. I upped MAX_TIMEOUT to 250; and the I2C bus seems to respond within that period. should it be necessary to increase it, I guess the m variable should be upgraded to uint16.